googleadk_tests/pyproject.toml
[project]
name = "googleadk-tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.23.0",
    "google-genai>=1.60.0",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.0",
]
